--- src/main/java/org/team1/nbe1_2_team01/domain/attendance/controller/dto/AttendanceIdRequest.java
@@ -1,7 +1,0 @@
-package org.team1.nbe1_2_team01.domain.attendance.controller.dto;
-
-public record AttendanceIdRequest(
-        Long id
-) {
-
-}
