+++ src/main/java/org/team1/nbe1_2_team01/domain/attendance/service/AttendanceUpdater.java
@@ -3,7 +3,7 @@ package org.team1.nbe1_2_team01.domain.attendance.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
-import org.team1.nbe1_2_team01.domain.attendance.controller.dto.AttendanceUpdateRequest;
+import org.team1.nbe1_2_team01.domain.attendance.service.dto.AttendanceUpdateRequest;
 import org.team1.nbe1_2_team01.domain.attendance.entity.Attendance;
 import org.team1.nbe1_2_team01.domain.attendance.service.port.AttendanceRepository;
