+++ src/main/java/org/team1/nbe1_2_team01/domain/attendance/service/dto/AttendanceDeleteRequest.java
@@ -1,0 +1,7 @@
+package org.team1.nbe1_2_team01.domain.attendance.service.dto;
+
+public record AttendanceDeleteRequest(
+        Long id
+) {
+
+}
