+++ src/main/java/org/team1/nbe1_2_team01/domain/attendance/service/dto/AttendanceUpdateRequest.java
@@ -1,4 +1,4 @@
-package org.team1.nbe1_2_team01.domain.attendance.controller.dto;
+package org.team1.nbe1_2_team01.domain.attendance.service.dto;
 
 import java.time.LocalDateTime;
 import lombok.Builder;
