+++ src/test/java/org/team1/nbe1_2_team01/domain/attendance/service/AttendanceServiceTest.java
@@ -14,8 +14,8 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.test.util.ReflectionTestUtils;
-import org.team1.nbe1_2_team01.domain.attendance.controller.dto.AttendanceCreateRequest;
-import org.team1.nbe1_2_team01.domain.attendance.controller.dto.AttendanceUpdateRequest;
+import org.team1.nbe1_2_team01.domain.attendance.service.dto.AttendanceCreateRequest;
+import org.team1.nbe1_2_team01.domain.attendance.service.dto.AttendanceUpdateRequest;
 import org.team1.nbe1_2_team01.domain.attendance.entity.Attendance;
 import org.team1.nbe1_2_team01.domain.attendance.entity.IssueType;
 import org.team1.nbe1_2_team01.domain.attendance.fake.AttendanceFakeRepository;
