+++ src/test/java/org/team1/nbe1_2_team01/domain/attendance/service/AttendanceUpdaterTest.java
@@ -7,8 +7,8 @@ import java.time.LocalDateTime;
 import java.time.LocalTime;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.team1.nbe1_2_team01.domain.attendance.controller.dto.AttendanceCreateRequest;
-import org.team1.nbe1_2_team01.domain.attendance.controller.dto.AttendanceUpdateRequest;
+import org.team1.nbe1_2_team01.domain.attendance.service.dto.AttendanceCreateRequest;
+import org.team1.nbe1_2_team01.domain.attendance.service.dto.AttendanceUpdateRequest;
 import org.team1.nbe1_2_team01.domain.attendance.entity.Attendance;
 import org.team1.nbe1_2_team01.domain.attendance.entity.IssueType;
 import org.team1.nbe1_2_team01.domain.attendance.fake.AttendanceFakeRepository;
